--- src/eru.WebApp/Assets/scss/common.scss
@@ -1,31 +1,0 @@
-﻿@import url('https://cdn.jsdelivr.net/npm/bulma@0.9.0/css/bulma.min.css');
-@import url('https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.0.0/animate.min.css');
-@import url('https://fonts.googleapis.com/css2?family=Quicksand:wght@300;400;500;600;700&display=swap');
-
-body {
-  display: flex;
-  min-height: 100vh;
-  flex-direction: column;
-}
-
-#wrapper {
-  flex: 1;
-}
-
-.chroma > path {
-  fill: black;
-  animation: chroma 8s linear infinite alternate both;
-}
-
-@keyframes chroma {
-  0% {
-    fill: #19dcea;
-  }
-
-  100% {
-    fill: #b22cff;
-  }
-}
-.title {
-  font-family: 'Quicksand', sans-serif;
-}
