--- src/eru.WebApp/Assets/scss/dashboard.scss
@@ -1,4 +1,0 @@
-﻿.center-table {
-  margin-left: auto;
-  margin-right: auto;
-}
