--- src/eru.WebApp/Assets/scss/upload.scss
@@ -1,1 +1,0 @@
-﻿
