+++ src/eru.WebApp/Pages/Admin.cshtml
@@ -1,21 +1,22 @@
 ﻿@page
 @using Microsoft.AspNetCore.Mvc.Localization
 @using Microsoft.Extensions.Configuration
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @inject IViewLocalizer Localizer
 @inject IConfiguration Configuration
 @{
-    Layout = "_Layout";
+    ViewData["title"] = Localizer["title"];
+    Layout = "Layouts/BaseLayout";
     //TODO Write how to setup SEQ in docs
     var seqServer = Configuration.GetSection("Serilog:WriteTo")?.GetChildren()?.Skip(2).FirstOrDefault()?.GetSection("Args:serverUrl")?.Value;
 }
 @section AddToHead
 {
-    <title>@Localizer["title"]</title>
-    <link rel="stylesheet" href="~/dist/dashboard.css" asp-append-version="true"/>
+    <link rel="stylesheet" href="~/dist/css/dashboard.min.css" asp-append-version="true"/>
 }
 @section AddToScripts
 {
-    <script src="~/dist/dashboard.js" asp-append-version="true"></script>
+    <script src="~/dist/js/dashboard.min.js" asp-append-version="true"></script>
 }
 <div id="wrapper">
     <nav class="navbar" role="navigation" aria-label="main navigation">
@@ -53,7 +54,7 @@
             <div class="tile is-child box has-text-centered">
                 <div class="field">
                     <label for="auto-refresh" class="has-text-weight-bold">Auto Refresh</label>
-                    <input id="auto-refresh" type="checkbox" name="auto-refresh-status" class="switch" onchange="localStorage.setItem('auto-refresh', this.checked)"/>
+                    <input id="auto-refresh" type="checkbox" name="auto-refresh-status" class="switch" onchange="toggleRefresh()"/>
                 </div>
             </div>
             <div class="tile is-child box">
@@ -106,7 +107,7 @@
                     </div>
                     <div class="field is-grouped">
                         <div class="control">
-                            <button id="create-class-button" class="button is-link">Create class</button>
+                            <button onclick="createClass()" class="button is-link">Create class</button>
                         </div>
                     </div>
                 </div>
