+++ src/eru.WebApp/Pages/Layouts/FormLayout.cshtml
@@ -1,0 +1,28 @@
+﻿@{
+    Layout = "Layouts/BaseLayout";
+}
+
+<svg class="chroma" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1440 320">
+    <path fill="#ff5500" fill-opacity="1" d="M0,192L180,288L360,192L540,192L720,256L900,160L1080,288L1260,64L1440,224L1440,0L1260,0L1080,0L900,0L720,0L540,0L360,0L180,0L0,0Z"></path>
+</svg>
+@section AddToHead
+{
+    @await RenderSectionAsync("AddToHead", false)
+    <link rel="stylesheet" href="dist/css/form.min.css"/>
+}
+@section AddToScripts
+{
+    @await RenderSectionAsync("AddToScripts", false)
+}
+<div id="wrapper">
+    <section class="hero">
+        <div class="container is-half">
+            <h1 class="is-size-1 title animate__animated animate__lightSpeedInRight">@ViewData["title"]</h1>
+        </div>
+    </section>
+    <section id="form-container" class="section">
+        <div class="container">
+            @RenderBody()
+        </div>
+    </section>
+</div>
