+++ src/eru.WebApp/Pages/Login.cshtml
@@ -3,49 +3,34 @@
 @model eru.WebApp.Pages.Login
 
 @{
-    Layout = "_Layout";
-}
-@section AddToHead
-{
-    <link rel="stylesheet" type="text/css" href="dist/login.css"/>
+    ViewData["title"] = "Login";
+    Layout = "Layouts/FormLayout";
 }
 
 @section AddToScripts
 {
-    <script src="dist/login.js"></script>
+    <script src="dist/js/login.min.js"></script>
 }
 
-<svg class="chroma" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 1440 320">
-    <path fill="#ff5500" fill-opacity="1" d="M0,192L180,288L360,192L540,192L720,256L900,160L1080,288L1260,64L1440,224L1440,0L1260,0L1080,0L900,0L720,0L540,0L360,0L180,0L0,0Z"></path>
-</svg>
-<div id="wrapper">
-    <section class="hero">
-        <div class="container is-half">
-            <h1 class="is-size-1 title animate__animated animate__lightSpeedInRight">Login</h1>
+<form method="post">
+    <div class="field">
+        <label class="label">Username</label>
+        <div class="control">
+            <input type="text" class="input" asp-for="Model.Username"/>
         </div>
-    </section>
-    <section id="form-container" class="columns">
-        <form method="post" class="container">
-            <div class="field">
-                <label class="label">Username</label>
-                <div class="control">
-                    <input type="text" class="input" asp-for="Model.Username"/>
-                </div>
-            </div>
-            <div class="field">
-                <label class="label">Password</label>
-                <div class="control">
-                    <input type="password" class="input" asp-for="Model.Password"/>
-                </div>
-            </div>
-            <div class="field is-grouped">
-                <div class="control">
-                    <input type="submit" class="button is-link" value="Submit">
-                </div>
-                <div class="control">
-                    <button type="button" id="cancel-button" class="button is-light">Cancel</button>
-                </div>
-            </div>
-        </form>
-    </section>
-</div>
+    </div>
+    <div class="field">
+        <label class="label">Password</label>
+        <div class="control">
+            <input type="password" class="input" asp-for="Model.Password"/>
+        </div>
+    </div>
+    <div class="field is-grouped">
+        <div class="control">
+            <input type="submit" class="button is-link" value="Submit">
+        </div>
+        <div class="control">
+            <button type="button" onclick="cancel()" class="button is-light">Cancel</button>
+        </div>
+    </div>
+</form>
