+++ src/eru.WebApp/eru.WebApp.csproj
@@ -10,6 +10,7 @@
     </ItemGroup>
   
     <ItemGroup>
+      <PackageReference Include="BuildBundlerMinifier" Version="3.2.449" />
       <PackageReference Include="Microsoft.AspNetCore.Mvc.Razor.RuntimeCompilation" Version="3.1.7" />
       <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="3.1.7">
         <PrivateAssets>all</PrivateAssets>
