+++ src/eru.WebApp/wwwroot/css/common.css
@@ -1,0 +1,13 @@
+﻿@import url('https://cdn.jsdelivr.net/npm/bulma@0.9.0/css/bulma.min.css');
+@import url('https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.0.0/animate.min.css');
+@import url('https://fonts.googleapis.com/css2?family=Quicksand:wght@300;400;500;600;700&display=swap');
+
+body {
+    display: flex;
+    min-height: 100vh;
+    flex-direction: column;
+}
+
+#wrapper {
+    flex: 1;
+}
