+++ src/eru.WebApp/wwwroot/css/dashboard.css
@@ -1,0 +1,4 @@
+﻿.center-table {
+    margin-left: auto;
+    margin-right: auto;
+}
