+++ src/eru.WebApp/wwwroot/css/form.css
@@ -1,0 +1,17 @@
+﻿.chroma > path {
+    fill: black;
+    animation: chroma 8s linear infinite alternate both;
+}
+
+@keyframes chroma {
+    0% {
+        fill: #19dcea;
+    }
+
+    100% {
+        fill: #b22cff;
+    }
+}
+.title {
+    font-family: 'Quicksand', sans-serif;
+}
