+++ src/eru.WebApp/wwwroot/js/login.js
@@ -1,0 +1,4 @@
+﻿function cancel() {
+    document.querySelector('#Model_Username').value = '';
+    document.querySelector('#Model_Password').value = '';
+}
